+++ .gitignore
@@ -33,7 +33,7 @@ MANIFEST
 *.manifest
 *.spec
 
-# Installer logs
+# Installer graph_logger
 pip-log.txt
 pip-delete-this-directory.txt
