+++ database/database.json
@@ -1,1 +1,1 @@
-{"109": {"wiggle_number": 19, "node_label": "NodeLabel", "created_at": 1666534101.384132, "updated_at": null, "belongings": {"uuid": "7e48f6ae-b25a-4634-91af-b1fb67b90ad9"}, "relations": null}}
+{"3794": {"wiggle_number": 3794, "node_label": "NodeLabel", "created_at": 1666534101.384132, "updated_at": null, "belongings": {"uuid": "7e48f6ae-b25a-4634-91af-b1fb67b90ad9"}, "relations": null}, "9070": {"wiggle_number": 9070, "node_label": "NodeLabel", "created_at": 1666534101.384132, "updated_at": null, "belongings": {"uuid": "7e48f6ae-b25a-4634-91af-b1fb67b90ad9"}, "relations": null}}
