+++ database/database.py
@@ -3,18 +3,21 @@ from json.decoder import JSONDecodeError
 
 from exceptions.database import NodeExistsError
 
+from graph_logger.graph_logger import graph_logger
+
 DATABASE_FILE_PATH = "database/database.json"
 
 
 def load_database(file_path: str) -> dict:
-
+    graph_logger.info("Atempting to loading database")
     try:
         with open(file_path, "r") as file_handle:
             database = load(file_handle)
-
+            graph_logger.info("Succesfully loaded database")
             return database
 
     except JSONDecodeError:
+        graph_logger.exception("Empty database, returning a new one")
         return {}
 
 
@@ -32,20 +35,47 @@ def add_item_to_database(file_path: str, item: dict):
     database.update(item)
 
     with open(file_path, "w") as file_handle:
+        graph_logger.info("writing to db")
         dump(database, file_handle, ensure_ascii=False)
+        graph_logger.info("Done")
 
 
-def wipe_database(file_path: str):
+def add_item_to_database_append(file_path: str, item: dict):
+    wiggle_number = None
+    for key in item:
+        # todo check agaisnts wiggle number state
+        wiggle_number = key
+        if str(wiggle_number) in {1, 2, 3}:
+            raise NodeExistsError(
+                f"Node {wiggle_number} already exisits did you mean to update"
+            )
 
-    with open(file_path, "w") as file_handle:
-        file_handle.write("")
+    with open(file_path, "a") as file_handle:
+        graph_logger.info(f"Writing Node {wiggle_number} to db")
+        dump(item, file_handle, ensure_ascii=False)
+        graph_logger.info(f"Succesfully wrote Node {wiggle_number} to db")
+
+
+def wipe_database(file_path: str, im_sure: bool = False):
+    if im_sure:
+        graph_logger.info("Dropping database")
+        with open(file_path, "w") as file_handle:
+            file_handle.write("")
+        return None
+    graph_logger.debug(f"Did not drop datbase as {im_sure=}")
 
 
 if __name__ == "__main__":
     x = load_database(DATABASE_FILE_PATH)
+
+    from random import randint
+
+    print(x)
+    wn = randint(0, 9999)
+
     i = {
-        109: {
-            "wiggle_number": 19,
+        wn: {
+            "wiggle_number": wn,
             "node_label": "NodeLabel",
             "created_at": 1666534101.384132,
             "updated_at": None,
@@ -54,7 +84,8 @@ if __name__ == "__main__":
         }
     }
     # update_database(DATABASE_FILE_PATH)
-
+    #
+    # add_item_to_database(DATABASE_FILE_PATH, i)
     add_item_to_database(DATABASE_FILE_PATH, i)
 
-    # wipe_database(DATABASE_FILE_PATH)
+    wipe_database(DATABASE_FILE_PATH)
