+++ state/wiggle_number.py
@@ -1,24 +1,28 @@
+from graph_logger.graph_logger import graph_logger
+
 WIGGLE_NUMBER_FILE_PATH = "state/wiggle_number.txt"
 
 
 def get_current_wiggle_number(file_path: str) -> int:
-
+    graph_logger.debug("Attempting getting Wiggle number")
     with open(file_path, "r") as file_handle:
         wiggle_number_string = file_handle.read()
 
     if wiggle_number_string:
+        graph_logger.debug("Wiggle number found, casting to int")
         wiggle_number = int(wiggle_number_string)
     else:
+        graph_logger.info("Wiggle number not set retuning 0")
         wiggle_number = 0
-
+    graph_logger.info(f"Succesfully got Wiggle number was {wiggle_number}")
     return wiggle_number
 
 
 def update_wiggle_number(file_path: str, wiggle_number: int) -> int:
-
+    graph_logger.debug("Updating state for wiggle number")
     with open(file_path, "w") as file_handle:
         file_handle.write(str(wiggle_number))
-
+    graph_logger.debug("Scuccesfully updated state for Wiggle number")
     return wiggle_number
