+++ state/wiggle_number.txt
@@ -1,1 +1,1 @@
-20
+50
